fix(tools+loader): center_x/y categorical→float 자동 캐스팅
- control center_x: category (문자열, 삼각형 버그 원인)
- lps     center_x: float64 (원래 정상)
- .astype(float).values 로 강제 캐스팅
- data_loader.load(): 좌표 컬럼 로드 시 자동 float 캐스팅
- 6장 leiden + 6장 cell_type 재생성, 뇌 해부 구조 육안 확인
- 시험 (2)-1 완료 (cortex 6층 · corpus callosum · striatum · hippocampus)

+++ task2_spatial_aging/src/tools.py
@@ -112,8 +112,8 @@ def plot_spatial_map(
     if adata.n_obs == 0:
         raise ValueError(f"조건에 맞는 세포 없음: {dataset}, slice={slice_id}")
 
-    x = adata.obs["center_x"].values
-    y = adata.obs["center_y"].values
+    x = adata.obs["center_x"].astype(float).values
+    y = adata.obs["center_y"].astype(float).values
     ct = adata.obs["cell_type"].astype(str).values
 
     all_ct = load(dataset).obs["cell_type"].astype(str).unique()
@@ -220,8 +220,8 @@ def plot_all_samples_grid(
         donor_mask = slice_data.obs["donor_id"].astype(str) == str(donor_id)
         sub = slice_data[donor_mask]
 
-        x = sub.obs["center_x"].values
-        y = sub.obs["center_y"].values
+        x = sub.obs["center_x"].astype(float).values
+        y = sub.obs["center_y"].astype(float).values
         cat = sub.obs[color_by].astype(str).values
         colors = [cat_to_color[c] for c in cat]
